refactor: InitService 패키지 이동
